1.0.18

+++ src/main/java/kiinse/plugins/ggo/gungalewiki/files/buttons/ButtonsData.java
@@ -6,12 +6,11 @@ import kiinse.plugins.ggo.gungaleapi.api.files.locale.PlayerLocale;
 import kiinse.plugins.ggo.gungalewiki.enums.Directory;
 import kiinse.plugins.ggo.gungalewiki.files.buttons.interfaces.FiltersButton;
 import kiinse.plugins.ggo.gungalewiki.files.buttons.interfaces.FiltersButtons;
-import org.bukkit.configuration.InvalidConfigurationException;
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.jetbrains.annotations.NotNull;
 
-import java.io.IOException;
 import java.util.HashMap;
+import java.util.logging.Level;
 
 public class ButtonsData extends FilesManager implements FiltersButtons {
 
@@ -24,14 +23,18 @@ public class ButtonsData extends FilesManager implements FiltersButtons {
     }
 
     @Override
-    public FiltersButtons load() throws IOException, InvalidConfigurationException {
-        if (isFileNotExists(Directory.BUTTONS) || isDirectoryEmpty(Directory.BUTTONS))
-            copyFile(Directory.BUTTONS);
-        for (var file : listFilesInDirectory(Directory.BUTTONS)) {
-            var yaml = new YamlConfiguration();
-            yaml.load(file);
-            buttons.put(Button.valueOf(file.getName().split("\\.")[0].toUpperCase()), new FiltersButtonImpl(yaml));
-        }
+    public FiltersButtons load() {
+        if (isFileNotExists(Directory.BUTTONS) || isDirectoryEmpty(Directory.BUTTONS)) copyFile(Directory.BUTTONS);
+        listFilesInDirectory(Directory.BUTTONS).forEach(file -> {
+            var fileName = file.getName().split("\\.")[0].toUpperCase();
+            try {
+                var yaml = new YamlConfiguration();
+                yaml.load(file);
+                buttons.put(Button.valueOf(fileName), new FiltersButtonImpl(yaml));
+            } catch (Exception e) {
+                plugin.sendLog(Level.WARNING, "Error on loading filter button '" + fileName + "'! Message: " + e.getMessage());
+            }
+        });
         return this;
     }
